Moved transactions template to test configurations in eventuate client jdbc modules.
