+++ eventuate-client-java-spring-jdbc/src/main/java/io/eventuate/javaclient/spring/jdbc/EmbeddedTestAggregateStoreConfiguration.java
@@ -13,18 +13,15 @@ import io.eventuate.javaclient.commonimpl.adapters.SyncToAsyncAggregateCrudAdapt
 import io.eventuate.javaclient.commonimpl.adapters.SyncToAsyncAggregateEventsAdapter;
 import io.eventuate.javaclient.eventhandling.exceptionhandling.EventuateClientScheduler;
 import io.eventuate.javaclient.jdbc.EventuateEmbeddedTestAggregateStore;
-import io.eventuate.javaclient.jdbc.JdkTimerBasedEventuateClientScheduler;
-import io.eventuate.javaclient.spring.common.EventuateCommonConfiguration;
 import io.eventuate.javaclient.jdbc.EventuateJdbcAccess;
 import io.eventuate.javaclient.jdbc.EventuateJdbcAccessImpl;
+import io.eventuate.javaclient.jdbc.JdkTimerBasedEventuateClientScheduler;
+import io.eventuate.javaclient.spring.common.EventuateCommonConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.transaction.support.TransactionTemplate;
 
-import javax.sql.DataSource;
 import java.util.Collections;
 
 @Configuration
@@ -35,11 +32,6 @@ public class EmbeddedTestAggregateStoreConfiguration {
   @Bean
   public EventuateDatabaseScriptSupplier eventuateCommonInMemoryScriptSupplierForEventuateLocal() {
     return () -> Collections.singletonList("eventuate-embedded-schema.sql");
-  }
-
-  @Bean
-  public TransactionTemplate transactionTemplate(DataSource dataSource) {
-    return new TransactionTemplate(new DataSourceTransactionManager(dataSource));
   }
