+++ eventuate-client-java-spring-jdbc/src/test/java/io/eventuate/javaclient/spring/jdbc/EventuateJdbcAccessImplTest.java
@@ -1,6 +1,7 @@
 package io.eventuate.javaclient.spring.jdbc;
 
-import io.eventuate.common.jdbc.*;
+import io.eventuate.common.jdbc.EventuateJdbcStatementExecutor;
+import io.eventuate.common.jdbc.EventuateSchema;
 import io.eventuate.javaclient.jdbc.EventuateJdbcAccess;
 import io.eventuate.javaclient.jdbc.common.tests.CommonEventuateJdbcAccessImplTest;
 import org.junit.Test;
